spec: add ability for plugins to return state to persist
Currently any state that the plugin needs to maintain has
to be serialized to disk by the plugin itself. To reduce
this burden, the current spec provides the DEL command
all the same information as the ADD command. This can be
tricky for the runtime to achieve, requiring it to make
a copy of network configuration. Since it has to do the
work of writing to disk, this change allows the plugin to
return extra state for it to persist. The DEL command
interface has been changed to pipe in previously returned
state. Therefore the DEL command no longer needs to receive
all the information that was provided to ADD as the plugin
can easily save it via the "state" mechanism.

Fixes #75