dhcp: add broadcast and sendClientID options
Broadcast is necessary for ipvlan plugin as
kernel will drop unicast replies since the slave
device has not yet been configured with an IP.

ClientID is "containerid/netname" and can help
with visibility into allocated IPs in DHCP server.
It is also important for ipvlan as all h/w addresses
are the same.

Fixes #17