Add libkv meta plugin
Allows to load the container network configuration from remote
stores supported by libkv.

It further adds the possibility to do per container network
configuration by reading an array of cni configurations from the store.
The first configuration is treated as the main config and it's
configuration result is returned to the caller.

A possible use case is with kubernetes. Kubernetes only reads the first
config from CNI and configures the pod with it. With this plugin we can
have individual pod network configurations by still having an identical
CNI configuration for the cluster managment (the first config in the
array).