Add identifier config property
With the `identifier` property every environment can make use of the
libkv plugin. It adds a generic way on how to find out the right storage
path in the key value store for a CNI run.

By default `config.basePath + cmdArgs.ContainerID` is used. When an
alternative identifer is provided the lookup path is determined with
`basePath + cmdArgs.Args[config.Identifier]`