pkg: ns: Fix GetNS() by assigning mounted flag if nspath mounted
In case you create a new netns, the handler returned has the mounted flag
assigned to true. If you want to get another handler with GetNS() because
you didn't store the first one, you won't have the mounted flag properly
assigned, meaning the Close() function won't work as expected because it
won't unmount the netns. This is a bug because the close is supposed to
remove the netns completely.

This patch checks if the network namespace path provided is currently mounted,
and in case it is, it will set the mounted flag to true.

Signed-off-by: Sebastien Boeuf <sebastien.boeuf@intel.com>