hwaddr: Generate MAC address for IPv6-only pods
This change adds logic to set the MAC (hardware) address for IPv6-only
pods.

Currently, the SetHWAddrByIP depends on an IPv4 IP address being
configured, and it sets the MAC address based on the IPv4 address by
wrapping the IPv4 address with a special, hard-coded 2-byte prefix:
    [0a:58] + [IPv4-address]
This works fine for the IPv4 case, since the 4 bytes of IPv4 address
is guaranteed to be unique. The IPv6-only case is not being handled.

What's being proposed with this change set is to modify SetHWAddrByIP
behavior as follows:
- IPv4-only or Dual-Stack (IPv4 + IPv6):
  No change in behavior, MAC address is based on IPv4 address:
    [0a:58] + [IPv4-address]
- IPv6-only:
  Generate and set a MAC address using a different hard-coded prefix
  to make the V6 MAC address space separate and discernible from the
  V4 MAC address space:
    [6a:58] + [four-random-octets]
  Here, the four random octets are derived by taking the last 4 bytes
  of the original MAC address as assigned by the OS when the intf/link
  is created. These are assumed to be random.

So a MAC address for an IPv6-only pod will depend upon the original
OS-assigned MAC address, and not the IPv6 address. The reason that the
MAC is not derived from the IPv6 address is that IPv6 addresses are
much larger than 32 bits (64 bits of interface ID + 16 or less bits of
subnet ID + 48 or more bits of routing prefix), and it's difficult, if
not impossible to come up with a scheme to select 32 bits from the
128 bytes of address that are guaranteed to be unique over a long
period of time as pods are created and destroyed. For example, when the
flannel plugin is used (which delegates to bridge plugin, and therefore
calls SetHWAddrByIP), each pod may get assigned a /64 subnet out of a
/48 range. In this case, the lower 64 bits of interface ID are not
unique, and the 16 bits of subnet ID will be getting recycled over time.

Another reason for not basing the MAC address on an IPv6 address is that
when IPv6 SLAAC support is added, the MAC address assignment will need
to be done before SLAAC addresses are generated by the pod, so we'll
need a mechanism for generating MACs independent of IPv6 addresses anyway.

This proposal is just one way to assign MAC addresses for the IPv6-only
case. Some alternatives that can be considered:
- Just keep all 6 bytes of the original OS-generated MAC address
- Use a shorter, 1-byte hard-coded prefix (e.g. 6a), allowing 5 bytes of
  randomness (less chance for MAC collisions).
- Develop a MAC address allocation pool mechanism, similar to what's
  done for IP address assignment with the host-local IPAM. This would
  involve a bunch more complexity.
- Hash the 128 bytes of IPv6 address into 32 bits. Some considerations
  for this approach:
      * The MAC address is already random, why add another randomization
        step?
      * This approach would be inconsistent with whatever we come up
        with for IPv6 SLAAC, where MAC needs to be set before IPv6
        address assignment (we'd have 2 ways V6 MACs are assigned).

Fixes #370