pkg/skel: return error if JSON config has no network name
SPEC.md has indicated that 'name' is required since at least mid-2015,
but that was not enforced in pkg/skel or our unit tests.  That could
lead to configs passing no network name and plugins like host-local
that require the name failing or causing odd behavior.

Since the name is required by the spec, require it in the code too.